python_ml/pytorch/mnist/Inference/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
onnxruntime==1.17.3
numpy==1.26.4
psutil==5.9.8
python-multipart
Pillow
